wip: start moving to eclair organizational structure
