+++ manta-crypto/src/constraint.rs
@@ -23,266 +23,16 @@
 //        vector by the wrong amount or in the wrong order.
 
 use crate::rand::{CryptoRng, RngCore};
-use core::{cmp, fmt::Debug, hash::Hash, marker::PhantomData, ops};
-use manta_util::{create_seal, seal};
 
-create_seal! {}
-
-/// Generic Derived Allocation Mode
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Derived;
-
-/// Public Allocation Mode
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Public;
-
-impl From<Derived> for Public {
-    #[inline]
-    fn from(d: Derived) -> Self {
-        let _ = d;
-        Self
-    }
-}
-
-/// Secret Allocation Mode
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct Secret;
-
-impl From<Derived> for Secret {
-    #[inline]
-    fn from(d: Derived) -> Self {
-        let _ = d;
-        Self
-    }
-}
-
-/// Constant Type Alias
-pub type Const<C, COM> = <C as Constant<COM>>::Type;
-
-/// Compiler Constant
-pub trait Constant<COM>
-where
-    COM: ?Sized,
-{
-    /// Underlying Type
-    type Type;
-
-    /// Allocates a new constant from `this` into the `compiler`.
-    fn new_constant(this: &Self::Type, compiler: &mut COM) -> Self;
-}
-
-impl<COM> Constant<COM> for ()
-where
-    COM: ?Sized,
-{
-    type Type = ();
-
-    #[inline]
-    fn new_constant(this: &Self::Type, compiler: &mut COM) -> Self {
-        let _ = (this, compiler);
-    }
-}
-
-impl<T, COM> Constant<COM> for PhantomData<T>
-where
-    COM: ?Sized,
-{
-    type Type = PhantomData<T>;
-
-    #[inline]
-    fn new_constant(this: &Self::Type, compiler: &mut COM) -> Self {
-        let _ = (this, compiler);
-        PhantomData
-    }
-}
-
-/// Variable Type Alias
-pub type Var<V, M, COM> = <V as Variable<M, COM>>::Type;
-
-/// Compiler Variable
-pub trait Variable<M, COM>
-where
-    COM: ?Sized,
-{
-    /// Underlying Type
-    type Type;
-
-    /// Allocates a new known value from `this` into the `compiler`.
-    fn new_known(this: &Self::Type, compiler: &mut COM) -> Self;
-
-    /// Allocates a new unknown value into the `compiler`.
-    fn new_unknown(compiler: &mut COM) -> Self;
-}
-
-impl<M, COM> Variable<M, COM> for ()
-where
-    COM: ?Sized,
-{
-    type Type = ();
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut COM) -> Self {
-        let _ = (this, compiler);
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut COM) -> Self {
-        let _ = compiler;
-    }
-}
-
-impl<T, M, COM> Variable<M, COM> for PhantomData<T>
-where
-    COM: ?Sized,
-{
-    type Type = PhantomData<T>;
-
-    #[inline]
-    fn new_known(this: &Self::Type, compiler: &mut COM) -> Self {
-        let _ = (this, compiler);
-        PhantomData
-    }
-
-    #[inline]
-    fn new_unknown(compiler: &mut COM) -> Self {
-        let _ = compiler;
-        PhantomData
-    }
-}
-
-/// Value Source Auto-Trait
-pub trait ValueSource<COM>
-where
-    COM: ?Sized,
-{
-    /// Allocates `self` as a constant in `compiler`.
-    #[inline]
-    fn as_constant<C>(&self, compiler: &mut COM) -> C
-    where
-        C: Constant<COM, Type = Self>,
-    {
-        C::new_constant(self, compiler)
-    }
-
-    /// Allocates `self` as a known value in `compiler`.
-    #[inline]
-    fn as_known<M, V>(&self, compiler: &mut COM) -> V
-    where
-        V: Variable<M, COM, Type = Self>,
-    {
-        V::new_known(self, compiler)
-    }
-
-    /// Allocates an unknown value of type `Self` into `compiler`.
-    #[inline]
-    fn as_unknown<M, V>(compiler: &mut COM) -> V
-    where
-        V: Variable<M, COM, Type = Self>,
-    {
-        V::new_unknown(compiler)
-    }
-}
-
-impl<COM, T> ValueSource<COM> for T where T: ?Sized {}
-
-/// Allocator Auto-Trait
-pub trait Allocator {
-    /// Allocates a constant with the given `value` into `self`.
-    #[inline]
-    fn allocate_constant<C>(&mut self, value: &C::Type) -> C
-    where
-        C: Constant<Self>,
-    {
-        C::new_constant(value, self)
-    }
-
-    /// Allocates a known variable with the given `value` into `self`.
-    #[inline]
-    fn allocate_known<M, V>(&mut self, value: &V::Type) -> V
-    where
-        V: Variable<M, Self>,
-    {
-        V::new_known(value, self)
-    }
-
-    /// Allocates an unknown variable into `self`.
-    #[inline]
-    fn allocate_unknown<M, V>(&mut self) -> V
-    where
-        V: Variable<M, Self>,
-    {
-        V::new_unknown(self)
-    }
-}
-
-impl<COM> Allocator for COM where COM: ?Sized {}
-
-/// Native Compiler Marker Trait
-///
-/// This `trait` is only implemented for `()`, the default native compiler.
-pub trait Native: sealed::Sealed {
-    /// Returns an instance of the native compiler.
-    fn compiler() -> Self;
-}
-
-seal! { () }
-
-impl Native for () {
-    #[inline]
-    fn compiler() -> Self {}
-}
-
-/// Boolean Type
-///
-/// This `trait` should be implemented for compilers that offer a boolean type equivalent to `bool`.
-pub trait HasBool {
-    /// Boolean Type
-    ///
-    /// This type should have a notion of `true` and `false` and negation.
-    type Bool: Not<Self, Output = Self::Bool>;
-}
-
-impl HasBool for () {
-    type Bool = bool;
-}
-
-/// Assertion
-pub trait Assert: HasBool {
-    /// Asserts that `b` reduces to `true`.
-    fn assert(&mut self, b: &Self::Bool);
-}
-
-impl Assert for () {
-    #[inline]
-    fn assert(&mut self, b: &Self::Bool) {
-        // TODO: USe `dbg!` macro here to get more info, but add a feature-flag for this.
-        assert!(b)
-    }
-}
-
-/// Equality Assertion
-pub trait AssertEq<T, Rhs = T>: Assert
-where
-    T: PartialEq<Rhs, Self>,
-{
-    /// Asserts that `lhs` and `rhs` are equal.
-    #[inline]
-    fn assert_eq(&mut self, lhs: &T, rhs: &Rhs) {
-        let are_equal = lhs.eq(rhs, self);
-        self.assert(&are_equal)
-    }
-}
-
-impl<T, Rhs> AssertEq<T, Rhs> for ()
-where
-    T: cmp::PartialEq<Rhs> + Debug,
-    Rhs: Debug,
-{
-    #[inline]
-    fn assert_eq(&mut self, lhs: &T, rhs: &Rhs) {
-        assert_eq!(lhs, rhs)
-    }
-}
+pub use crate::eclair::{
+    alloc::{
+        mode::{self, Derived, Public, Secret},
+        Allocate, Allocator, Const, Constant, Var, Variable,
+    },
+    cmp::{Eq, HasBool, PartialEq},
+    ops::{Add, Assert, AssertEq, Not, Sub},
+    Native,
+};
 
 /// Constraint System
 pub trait ConstraintSystem {
@@ -369,103 +119,6 @@ pub trait ConstraintSystem {
     */
 }
 
-///
-pub trait Eq<COM = ()>: PartialEq<Self, COM>
-where
-    COM: HasBool,
-{
-}
-
-impl<T> Eq for T where T: cmp::Eq {}
-
-///
-pub trait Not<COM = ()>
-where
-    COM: ?Sized,
-{
-    ///
-    type Output;
-
-    ///
-    fn not(self, compiler: &mut COM) -> Self::Output;
-}
-
-impl<T> Not for T
-where
-    T: ops::Not,
-{
-    type Output = T::Output;
-
-    #[inline]
-    fn not(self, _: &mut ()) -> Self::Output {
-        self.not()
-    }
-}
-
-///
-pub trait PartialEq<Rhs = Self, COM = ()>
-where
-    Rhs: ?Sized,
-    COM: HasBool + ?Sized,
-{
-    ///
-    fn eq(&self, rhs: &Rhs, compiler: &mut COM) -> COM::Bool;
-
-    ///
-    #[inline]
-    fn ne(&self, other: &Rhs, compiler: &mut COM) -> COM::Bool {
-        self.eq(other, compiler).not(compiler)
-    }
-
-    /* TODO:
-       /// Asserts that `lhs` and `rhs` are equal.
-       #[inline]
-       fn assert_eq(lhs: &Self, rhs: &Self, compiler: &mut COM) {
-           let boolean = lhs.eq(rhs, compiler);
-           compiler.assert(boolean);
-       }
-
-       /// Asserts that all the elements in `iter` are equal to some `base` element.
-       #[inline]
-       fn assert_all_eq_to_base<'t, I>(base: &'t Self, iter: I, compiler: &mut COM)
-       where
-           I: IntoIterator<Item = &'t Self>,
-       {
-           for item in iter {
-               Self::assert_eq(base, item, compiler);
-           }
-       }
-
-       /// Asserts that all the elements in `iter` are equal.
-       #[inline]
-       fn assert_all_eq<'t, I>(iter: I, compiler: &mut COM)
-       where
-           Self: 't,
-           I: IntoIterator<Item = &'t Self>,
-       {
-           let mut iter = iter.into_iter();
-           if let Some(base) = iter.next() {
-               Self::assert_all_eq_to_base(base, iter, compiler);
-           }
-       }
-    */
-}
-
-impl<T, Rhs> PartialEq<Rhs> for T
-where
-    T: cmp::PartialEq<Rhs>,
-{
-    #[inline]
-    fn eq(&self, rhs: &Rhs, _: &mut ()) -> bool {
-        self.eq(rhs)
-    }
-
-    #[inline]
-    fn ne(&self, rhs: &Rhs, _: &mut ()) -> bool {
-        self.ne(rhs)
-    }
-}
-
 /* TODO:
 /// Conditional Selection
 pub trait ConditionalSelect<COM>
@@ -499,44 +152,6 @@ where
     }
 }
 */
-
-/// Addition
-pub trait Add<COM>
-where
-    COM: ?Sized,
-{
-    /// Adds `lhs` and `rhs` inside of `compiler`.
-    fn add(lhs: Self, rhs: Self, compiler: &mut COM) -> Self;
-}
-
-impl<T> Add<()> for T
-where
-    T: ops::Add<Output = T>,
-{
-    #[inline]
-    fn add(lhs: Self, rhs: Self, _: &mut ()) -> Self {
-        lhs.add(rhs)
-    }
-}
-
-/// Subtraction
-pub trait Sub<COM>
-where
-    COM: ?Sized,
-{
-    /// Subtracts `rhs` from `lhs` inside of `compiler`.
-    fn sub(lhs: Self, rhs: Self, compiler: &mut COM) -> Self;
-}
-
-impl<T> Sub<()> for T
-where
-    T: ops::Sub<Output = T>,
-{
-    #[inline]
-    fn sub(lhs: Self, rhs: Self, _: &mut ()) -> Self {
-        lhs.sub(rhs)
-    }
-}
 
 /// Proof System
 pub trait ProofSystem {
@@ -611,9 +226,10 @@ where
 
 /// Constraint System Measurement
 pub mod measure {
-    use super::*;
+    use super::mode::{Constant, Public, Secret};
     use core::{
-        fmt::Display,
+        fmt::{Debug, Display},
+        hash::Hash,
         ops::{Add, AddAssign, Deref, DerefMut},
     };
 
@@ -623,37 +239,27 @@ pub mod measure {
     #[cfg(feature = "serde")]
     use manta_util::serde::{Deserialize, Serialize};
 
+    /// Variable Counting
+    pub trait Count<M> {
+        /// Returns the number of variables of the given mode `M` are stored in `self`.
+        fn count(&self) -> Option<usize> {
+            None
+        }
+    }
+
     /// Constraint System Measurement
-    pub trait Measure {
+    pub trait Measure: Count<Constant> + Count<Public> + Count<Secret> {
         /// Returns the number of constraints stored in `self`.
         fn constraint_count(&self) -> usize;
-
-        /// Returns the number of allocated constants.
-        #[inline]
-        fn constant_count(&self) -> Option<usize> {
-            None
-        }
-
-        /// Returns the number of allocated public variables.
-        #[inline]
-        fn public_variable_count(&self) -> Option<usize> {
-            None
-        }
-
-        /// Returns the number of allocated secret variables.
-        #[inline]
-        fn secret_variable_count(&self) -> Option<usize> {
-            None
-        }
 
         /// Returns a [`Size`] with the number of constraints and variables of each kind.
         #[inline]
         fn measure(&self) -> Size {
             Size {
                 constraint_count: self.constraint_count(),
-                constant_count: self.constant_count(),
-                public_variable_count: self.public_variable_count(),
-                secret_variable_count: self.secret_variable_count(),
+                constant_count: Count::<Constant>::count(self),
+                public_variable_count: Count::<Public>::count(self),
+                secret_variable_count: Count::<Secret>::count(self),
             }
         }
