+++ manta-crypto/src/eclair/measure.rs
@@ -14,27 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Cryptographic Primitives Library
-
-#![cfg_attr(not(feature = "std"), no_std)]
-#![cfg_attr(doc_cfg, feature(doc_cfg))]
-#![forbid(rustdoc::broken_intra_doc_links)]
-#![forbid(missing_docs)]
-
-#[cfg(feature = "alloc")]
-extern crate alloc;
-
-pub mod accumulator;
-pub mod algebra;
-pub mod commitment;
-pub mod constraint;
-pub mod ecc;
-pub mod eclair;
-// TODO: pub mod encryption;
-pub mod hash;
-pub mod key;
-// TODO: pub mod merkle_tree;
-pub mod password;
-pub mod permutation;
-pub mod rand;
-pub mod signature;
+//! Compiler Measurement and Introspection
